update medicas app

+++ ethGui-MEDICAS/src/customtimer.h
@@ -34,7 +34,7 @@ public:
             endTime = m_EndTime;
         }
 
-        return std::chrono::duration_cast<std::chrono::milliseconds>(endTime - m_StartTime).count();
+        return (double)std::chrono::duration_cast<std::chrono::milliseconds>(endTime - m_StartTime).count();
     }
